Smaller refactor for same effect, per Brendan's review.
