+++ pwiz_tools/Skyline/TestUtil/TestUtil.csproj
@@ -146,7 +146,6 @@
     <Compile Include="AuditLogUtil.cs" />
     <Compile Include="CheckDocumentState.cs" />
     <Compile Include="CheckReportCompatibility.cs" />
-    <Compile Include="DifferenceFinder.cs" />
     <Compile Include="ExampleText.cs" />
     <Compile Include="ExtensionTestContext.cs" />
     <Compile Include="FormSeen.cs" />
